split shadow map logic into lighting_mesh_renderer submodule
